Databases are big: allow us to know that
A user had erroring on a parse of a BLAST output use to this being
over max int. We want to be able to use this on 32 bit machines, so
it is reasonable to force this to int64.

+++ blast/output.go
@@ -234,7 +234,7 @@ type Parameters struct {
 // A Statistics holds the search and database statistics for a Blast result.
 type Statistics struct {
 	DbNum    int     `xml:"Statistics_db-num"`    // Statistics_db-num
-	DbLen    int     `xml:"Statistics_db-len"`    // Statistics_db-len
+	DbLen    int64   `xml:"Statistics_db-len"`    // Statistics_db-len
 	HspLen   int     `xml:"Statistics_hsp-len"`   // Statistics_hsp-len
 	EffSpace float64 `xml:"Statistics_eff-space"` // Statistics_eff-space
 	Kappa    float64 `xml:"Statistics_kappa"`     // Statistics_kappa
